Do Some works

+++ src/main/kotlin/com/otakusaikou/simplerss/model/Subscribers.kt
@@ -3,5 +3,5 @@ package com.otakusaikou.simplerss.model
 import org.jetbrains.exposed.dao.id.IntIdTable
 
 object Subscribers : IntIdTable() {
-    val qqNumber = long("qq_number")
+    val qqNumber = integer("qq_number")
 }
